requirements.txt
pydantic~=2.10.6
openai~=1.66.3
tenacity~=9.0.0
pyyaml~=6.0.2
loguru~=0.7.3
numpy
datasets~=3.4.1
fastapi~=0.115.11
tiktoken~=0.9.0

html2text~=2024.2.26
gymnasium~=1.1.1
pillow~=11.1.0
browsergym~=0.13.3
uvicorn~=0.34.0
unidiff~=0.7.5
browser-use~=0.1.40
googlesearch-python~=1.3.0
baidusearch~=1.0.3
duckduckgo_search~=7.5.3

aiofiles~=24.1.0
pydantic_core~=2.27.2
colorama~=0.4.6
playwright~=1.51.0

docker~=7.1.0
pytest~=8.3.5
pytest-asyncio~=0.25.3

mcp~=1.5.0
httpx>=0.27.0
tomli>=2.0.0

boto3~=1.37.18

requests~=2.32.3
beautifulsoup4~=4.13.3

huggingface-hub~=0.29.2
setuptools~=75.8.0

# Bug Hunting and Security Testing Dependencies
python-nmap~=0.7.1
scapy~=2.6.1
python-whois~=0.9.4
dnspython~=2.7.0
paramiko~=3.5.0
pwntools~=4.13.1
urllib3~=2.3.0
lxml~=5.3.0
selenium~=4.28.0
shodan~=1.31.0
censys~=2.2.15
netaddr~=1.3.0

# Reporting and templating
jinja2~=3.1.4
markdown~=3.7.0
reportlab~=4.2.5

# Cryptography and security
cryptography~=44.0.0
pycryptodome~=3.21.0

# Web testing and fuzzing
websockets~=14.1

# Additional security tools integration
# Note: Some tools like nmap, nikto, gobuster, etc. need to be installed separately on Kali Linux

# Async and concurrency
asyncio-throttle~=1.0.2

# Configuration and CLI enhancements
typer~=0.15.1
rich~=13.9.4
colorlog~=6.9.0

# Database support for results storage
sqlalchemy~=2.0.36

# Development and testing
mypy~=1.14.1
black~=24.12.0
flake8~=7.1.1
